--- src/Pages/daeheelee/Main/Main.scss
@@ -1,366 +1,0 @@
-@import "Reset.css";
-.instamain{
-	nav {
-			display: flex;
-			position: fixed;
-			justify-content: center;
-			top:0px;
-			height: 7vh;
-			width:100%;
-			background-color: rgb(255, 255, 255);
-			border-bottom: solid 1px rgb(199,199,199);
-	}
-
-	.instamain {
-			display: flex;
-			justify-content: center;
-	}
-
-	.left {
-			display: flex;
-			position: relative;
-			align-items: center;
-	}
-
-	nav .rightimg{
-			width: 30px;
-			height: 30px;
-	}
-
-	.search{
-			width: 300px;
-			padding-left:3%;
-			background-color: rgb(250, 250, 250);
-			margin: 0px 180px;
-			border:solid 1px rgb(219,219,219);
-			outline:none;
-			font-size:16px;
-	}
-
-	.center{
-			display: flex;
-			justify-content: center;
-	}
-
-	.logo{
-			margin-left:20px;
-			padding-left: 20px;
-			border-left: solid 2px gray;
-			font-family: 'Lobster', cursive;
-			font-size: 30px;
-	}
-
-	.left img{
-			position: relative;
-			top:10px;
-			width: 120px;
-			height: 42px; 
-			padding-left:15px;
-			margin-left:15px;
-			border-left:solid 2px rgb(141, 140, 140);
-	}
-
-	.right{
-			display: flex;
-			justify-content: space-around;
-			margin:auto 0px;
-			width:17vh;
-	}
-
-	.center .search{
-			position: relative;
-			top:2.2vh;
-			height: 30px;
-			width:200px;
-	}
-
-	.main{
-			display: flex;
-			justify-content: center;
-			width: 100%;
-			height: 1100px;
-			margin-top:70px;
-			background-color: rgb(250,250,250);
-	}
-
-	.feeds{
-			width:610px;
-	}
-
-	.mainRight{
-			right:10px;
-			width:330px;
-			height: 600px;
-			margin-top:20px;
-			margin-left: 1vh;
-	}
-
-	.story{
-			height: 100px;
-			width:610px;
-			margin: 20px 0px;
-			background-color: rgb(255, 255, 255);
-			border:solid 1px rgb(219,219,219);
-	}
-
-	section{
-			width:610px;
-			background-color: rgb(255, 255, 255);
-			border:solid 1px rgb(219,219,219);
-	}
-
-	.header{
-			display: flex;
-			justify-content: space-between;
-			height:60px;
-	}
-
-	.header img{
-			width:45px;
-			height:45px;
-			border-radius: 40px;
-	}
-
-	.header2{
-			margin-top: 10px;
-	}
-
-	.dotIcon{
-			margin:auto 10px;
-			color:rgb(110,110,110);
-	}
-
-	.flex{
-			margin:6px;
-			margin-left: 10px;
-			display: flex;
-			justify-content: space-between;
-			width:185px;
-	}
-
-	.header2 p{
-			font-size:15px;
-	}
-
-	.imageSection img{
-			width:607.5px;
-			height:565px;
-			border:solid 1px black;
-	}
-
-	.iconline{
-			display: flex;
-			justify-content: space-between;
-			margin: 15px;
-			height:35px;
-			margin-top:0px;
-	}
-
-	.iconBox{
-			display: flex;
-			justify-content: space-between;
-			width:150px;
-	}
-
-	.like p{
-			position: relative;
-			margin-left:20px;
-			top:-10px;
-	}
-
-	.text{
-			position: relative;
-			top:-5px;
-			margin:0 20px;
-	}
-
-	.name{
-			font-weight: bold;
-			margin-right:5px;
-	}
-
-	.comments{
-			position: relative;
-			top:-5px;
-			margin:0 20px;
-	}
-
-	.textremove{
-			position: absolute;
-			right:30px;
-	}
-
-	.textremove{
-			border:none;
-			background-color: white;
-			color: rgb(219,219,219);
-			outline: none;
-	}
-
-	.textremove:hover{
-			color:rgb(58, 58, 58);
-	}
-
-	.divHeart{ 
-			display:inline;
-			position: absolute;
-			right:3px;
-			color: rgb(219,219,219);
-	}
-
-	.inputComment{
-			border-top:solid 1px rgb(230,230,230);
-			display: flex;
-			position: relative;
-			bottom: -2px;
-			height: 60px;
-	}
-
-	.smile{
-			padding: 15px;
-	}
-
-	.post{
-			width:470px;
-	}
-
-	.post input{
-			width: 100%;
-			height: 60%;
-			margin-top: 10px;
-			border:none;
-			outline:none;
-	}
-
-	.inputComment button{
-			width: 70px;
-			height: 55px;
-			border:none;
-			background-color: rgb(255, 255, 255);
-			color: rgb(191,224,253);
-			outline:none;
-	}
-
-	.rightProfile img{
-			border-radius: 40px;
-			width: 60px;
-			height: 60px;
-	}
-
-	.rightProfile{
-			display: flex;
-			justify-content: space-around;
-			align-items: center;
-			height: 90px;
-	}
-
-	.nameProfile{
-			margin-right: 60px;
-	}
-
-	.mainName{
-			font-weight: bold;
-			font-size: 15px;
-	}
-
-	.subname{
-			color:gray;
-			font-size: 15px;
-	}
-
-	.change p{
-			color: rgb(50,159,248);
-			font-size: 13px;
-	}
-
-	.commendtext{
-			display: flex;
-			justify-content: space-around;
-	}
-
-	.commendtext-right{
-			margin-left: 110px;
-	}
-
-	.commendtext span{
-			font-size:14px;
-	}
-
-	.commendtext-left{
-			font-weight: 600;
-			color:gray;
-	}
-
-	.num1Img img{
-			height: 40px;
-			width: 40px;
-			margin-top: 3px;
-			border-radius: 40px;
-	}
-
-	.rightNum1{
-			display: flex;
-			justify-content: space-around;
-			align-items: center;
-			margin: 5px 0;
-	}
-
-	.num1Name{
-			margin-right: 70px;
-	}
-
-	.num1Subname{
-			color:gray;
-			font-size: 13px;
-	}
-
-	.follow{
-			color:rgb(0,150,246);
-			font-size: 13px;
-			font-weight: 600;
-	}
-
-
-	.list{
-			margin: 30px 10px;
-	}
-
-	.footer1,.footer2{
-			display: flex;
-			margin-bottom: 5px;
-			color: rgb(199,199,199);
-			font-size: 11px;
-	}
-
-	li{
-			list-style: none;
-	}
-
-	.list p{
-			color: rgb(199,199,199);
-			font-size: 11px;
-	}
-
-	.wecodeStory1 img{
-			width: 60px;
-			height: 60px;
-			border-radius: 40px;
-	}
-
-	.wecodeStory1_name p{
-			padding:0 5px;
-			font-size: 14px;
-	}
-
-	.story{
-			padding:10px 20px;
-			width: 570px;
-			height: 80px;
-	}
-
-	.wecodeStory1{
-			display: block;
-			width:70px;
-			padding:0px 1px 0px 5px;
-	}
-}
