+++ src/Styles/Nav.scss
@@ -1,0 +1,54 @@
+nav {
+  display: flex;
+  position: fixed;
+  justify-content: center;
+  top: 0px;
+  height: 7vh;
+  width: 100%;
+  background-color: rgb(255, 255, 255);
+  border-bottom: solid 1px rgb(199, 199, 199);
+
+  .left {
+    display: flex;
+    position: relative;
+    align-items: center;
+
+    .logo {
+      margin-left: 20px;
+      padding-left: 20px;
+      border-left: solid 2px gray;
+      font-family: "Lobster", cursive;
+      font-size: 30px;
+    }
+  }
+
+  .rightimg {
+    width: 30px;
+    height: 30px;
+  }
+
+  .center {
+    display: flex;
+    justify-content: center;
+
+    .search {
+      position: relative;
+      top: 2.2vh;
+      height: 30px;
+      width: 200px;
+      padding-left: 3%;
+      margin: 0px 180px;
+      border: solid 1px rgb(219, 219, 219);
+      background-color: rgb(250, 250, 250);
+      outline: none;
+      font-size: 16px;
+    }
+  }
+
+  .right {
+    display: flex;
+    justify-content: space-around;
+    margin: auto 0px;
+    width: 17vh;
+  }
+}
